LayoutTests/accessibility/custom-elements/shadow-element-text-expected.txt

This test ensures that shadow dom elements return the right text.

Button1 text: 	AXTitle: Button One
	AXDescription:
	AXHelp:
Button2 text: 	AXTitle: Nested Button Two
	AXDescription:
	AXHelp:

PASS successfullyParsed is true

TEST COMPLETE
Button One
Nested Button Two
